fix: allow reads from the final addressable page (#140)
HTTP range validation rejected page 65535, excluding the final 16 MiB of
the 1 TiB address space. Include that page in the exclusive upper bound
and use the range iterator's `usize` length so requests spanning all
65,536 pages cannot overflow page-count accounting.

Regression coverage checks HTTP acceptance and rejection at the limit,
final-page mapping, and full-range metrics without downloading object
data. The new tests reproduced both the original rejection and the
overflow exposed by fixing only the bound.

Validation: 100 tests passed, 1 ignored with `cargo nextest run --locked
--all-features`; nightly formatting, Clippy across all features and
targets with warnings denied, `cargo deny check`, and `git diff --check`
passed.

Closes #127

+++ src/service/metrics.rs
@@ -113,7 +113,7 @@ pub fn fetch_request_bytes(kind: &ObjectKind, bytes: u64) {
         .observe(bytes as f64);
 }
 
-pub fn fetch_request_pages(kind: &ObjectKind, pages: u16) {
+pub fn fetch_request_pages(kind: &ObjectKind, pages: usize) {
     static HISTOGRAM: LazyLock<HistogramVec> = LazyLock::new(|| {
         register_histogram_vec!(
             "cachey_fetch_request_pages",
@@ -126,7 +126,7 @@ pub fn fetch_request_pages(kind: &ObjectKind, pages: u16) {
 
     HISTOGRAM
         .with_label_values(&[&**kind])
-        .observe(f64::from(pages));
+        .observe(pages as f64);
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
